refactor(StampType): 패키지 변경
- 기존 패키지 내부에 한 개의 클래스만 존재했기에 model 클래스에 추가
