+++ src/main/java/com/example/server/domain/stamp/persistence/StampRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.server.domain.stamp.model.Stamp;
-import com.example.server.domain.stamp.model.constants.StampType;
+import com.example.server.domain.stamp.model.StampType;
 
 @Repository
 @Transactional
